+++ configs/eslint-config/CHANGELOG.md
@@ -1,3 +1,10 @@
+`2023-07-27` - `0.0.6`
+
+- Match version with `eslint-config-react`
+- Updated dependencies to latest version
+- Disabled `@typescript-eslint/no-var-requires` in `gulpfile`
+- Disabled `no-param-reassign` in `gulpfile`
+
 `2023-07-27` - `0.0.2`
 
 - Moved unnecessary dependencies to devDependencies